VERY silly workaround for delocalization
In order to get a file output in WDL, you either have to know its full name before runtime or you can use glob() if you only know some of its name. We don't want to use glob() because glob() creates a folder with an unpredictable name which makes dealing with outputs manually (which we have to do as subsequent runs need to take in some files form previous runs manually) a pain. So to avoid glob(), we gotta know the name before runtime. But it's a function of the date because versioning is important. So we gotta set date. But we don't want to set it manually because we want the workflow to be automated (even though it isn't, and requires manual inputs, but those inputs could be handled automatically in theory if they do not have randomized folder names, ie, are not output via glob()). Therefore, the upstream matOptimize task now has two lines that call coreutils date, write it to a file named "date", and then read that file as a string in order to set a WDL variable for the date. Every time I write a workaround like this, I feel wiser -- and yet, at the same time, less smart.