Add x-distance ID failsafe, improve decimateds
Adds a renaming failsafe if we somehow end up with clusters at different distances with the same name, and a failsafe to the failsafe if cluster names aren't numeric, and a failsafe to the failsafe to the failsafe. Currently this ONLY runs in the persistent ID case, so don't mess up if you're running ad-hoc!

Decimated clusters are now explicitly flagged as such.

Decimated clusters don't get distance of ERROR!!!! anymore

Decimated clusters won't throw an error at the end anymore