Add comment, remove diff_datestamps
diff_datestamps was part of a proposed workaround for samples having the same sample name. It's better to use a deduplicated entity_id column in Terra instead via rename_samples, or just don't pass in samples with the same sample name.

Adding a comment can be useful for Terra data tables.